# mapaviajero-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Travel App API",
        "description": "Documentación de la API Travel App",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/idiomas": {
            "get": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Listar todos los idiomas",
                "description": "Obtiene la lista de todos los idiomas disponibles. Solo accesible para usuarios registrados.",
                "operationId": "70f14e34300a95dca5745cf7c529a94d",
                "responses": {
                    "200": {
                        "description": "Lista de idiomas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Idioma"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Crear un nuevo idioma",
                "description": "Solo accesible para administradores y superadministradores. Permite crear y añadir un nuevo idioma a la Base de Datos.",
                "operationId": "74e627db87abee90704ccee648b8b163",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IdiomaDB"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Idioma creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo los administradores pueden guardar un nuevo idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/idiomas/{id}": {
            "get": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Mostrar un idioma específico (por id)",
                "description": "Devuelve un idioma específico por su ID.",
                "operationId": "df819a81cb45d3f299d96f262f3fb3ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del idioma a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Idioma encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Idioma no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Idioma no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Actualizar un idioma",
                "description": "Permite modificar los datos de un idioma. Solo accesible para administradores y superadministradores.",
                "operationId": "fb0c6c6e714c9d957933d31d324377f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del idioma a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IdiomaDB"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Idioma actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede editar un idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Idioma no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Idioma no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ya existe un idioma con ese nombre en nuestra BBDD."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Eliminar un idioma",
                "description": "Permite eliminar un idioma. Solo accesible para administradores y superadministradores.",
                "operationId": "1aea638ff9bd9fff32197ac01e5e625b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del idioma a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Idioma borrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Idioma borrado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Solo el administrador puede borrar un idioma"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Idioma no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Idioma no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Idioma": {
                "title": "Idioma (Recurso API)",
                "properties": {
                    "id": {
                        "description": "ID del idioma",
                        "type": "integer",
                        "example": 1
                    },
                    "iso_639_1": {
                        "description": "Código ISO 639-1 del idioma",
                        "type": "string",
                        "example": "es"
                    },
                    "nombre": {
                        "description": "Nombre del idioma",
                        "type": "string",
                        "example": "Español"
                    }
                },
                "type": "object"
            },
...
            "IdiomaDB": {
                "title": "Idioma (BBDD)",
                "required": [
                    "nombre",
                    "iso_639_1"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del idioma",
                        "type": "string",
                        "example": "Catalán"
                    },
                    "iso_639_1": {
                        "description": "Código ISO 639-1 del idioma",
                        "type": "string",
                        "example": "ca"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}